requirements.txt
django
requests
psycopg2
schedule
Pillow
apscheduler
python-dateutil
djangorestframework